+++ sdk/client/src/types.ts
@@ -28,6 +28,7 @@ export interface SpendBuildParams {
   note: ShieldedNote;
   witness: MerkleWitness;
   nullifierSecret: Hex;
+  changeNullifierSecret?: Hex;
   merchantPubKey: Hex;
   merchantRho?: Hex;
   merchantAddress: Hex;
@@ -40,6 +41,7 @@ export interface SpendBuildParams {
 export interface SpendProofBundle {
   response: ShieldedPaymentResponse;
   changeNote: ShieldedNote;
+  changeNullifierSecret: Hex;
   merchantRho: Hex;
 }
 
@@ -59,4 +61,5 @@ export interface Prepared402Payment {
   headers: Headers;
   response: ShieldedPaymentResponse;
   changeNote: ShieldedNote;
+  changeNullifierSecret: Hex;
 }